Implement update_hourly_prices

+++ lib/osrs_ge_tracker/ge/hourly_price.ex
@@ -3,12 +3,12 @@ defmodule OsrsGeTracker.GE.HourlyPrice do
   import Ecto.Changeset
 
   schema "hourly_prices" do
-    field :buy_avg, :integer
-    field :buy_qty, :integer
-    field :overall_avg, :integer
-    field :overall_qty, :integer
-    field :sell_avg, :integer
-    field :sell_qty, :integer
+    field :buy_avg, :float
+    field :buy_qty, :float
+    field :overall_avg, :float
+    field :overall_qty, :float
+    field :sell_avg, :float
+    field :sell_qty, :float
     belongs_to :item, OsrsGeTracker.GE.Item
 
     timestamps(updated_at: false)
