+++ priv/repo/migrations/20190524003920_create_minutely_prices.exs
@@ -4,12 +4,12 @@ defmodule OsrsGeTracker.Repo.Migrations.CreateMinutelyPrices do
   def change do
     create table(:minutely_prices) do
       add :item_id, references(:items)
-      add :buy_avg, :integer
-      add :sell_avg, :integer
-      add :overall_avg, :integer
-      add :buy_qty, :integer
-      add :sell_qty, :integer
-      add :overall_qty, :integer
+      add :buy_avg, :float
+      add :sell_avg, :float
+      add :overall_avg, :float
+      add :buy_qty, :float
+      add :sell_qty, :float
+      add :overall_qty, :float
 
       timestamps(updated_at: false)
     end
